Remove pooling implementation
Pooling seems to cause problems with Tomcat 8. Disabling it has only a minor
effect on performance. This could be revisited in the future to further
improve processing but at the very least provides a considerable advantage
over the existing Hydra characterization method of invoking FITS on the
command line hundreds (or thousands) of times